+++ feed/feed.php
@@ -168,7 +168,7 @@ class feed
 				[
 					'FROM'	=> [$this->contests_table => 'c'],
 					'ON'	=> 'i.image_album_id = c.contest_album_id
-						AND c.contest_marked = ' . (int) \phpbbgallery\core\block::IN_CONTEST,
+						AND c.contest_marked <> ' . (int) \phpbbgallery\core\block::NO_CONTEST,
 				],
 			],
 			'WHERE'		=> $where,